Remove some more of the magic numbers
Replaced magic numbers with super generic FAILURE SUCCESS flags, its not much
better but its a start. Removed the string to uint function as it is potentially
unsafe with overflows and negative numbers. Instead use signed data types in all
but explicitly safe cases. Added a new string -> long function for use with
large numbers like cpu frequencies in MHz on very fast CPUs

+++ app/src/psfreq_cpu.h
@@ -36,14 +36,14 @@
 
 typedef struct psfreq_cpu_type {
         unsigned char has_pstate;
-        unsigned int cpu_num;
+        int cpu_num;
         char **vector_scaling_max_freq;
         char **vector_scaling_min_freq;
         char **vector_scaling_governor;
-        unsigned int cpuinfo_min_freq;
-        unsigned int cpuinfo_max_freq;
-        unsigned int scaling_min_freq;
-        unsigned int scaling_max_freq;
+        long cpuinfo_min_freq;
+        long cpuinfo_max_freq;
+        long scaling_min_freq;
+        long scaling_max_freq;
         char pst_turbo;
         char pst_max;
         char pst_min;
